flatmap-filter optimization
`collect` methods on flat-map with or without filter are optimized. In a 32 logical thread cpu, benchmarks are very promising: ~20 times faster than rayon. rayon seems to underperform sequential execution for some reason. orx-parallel, on the other hand, provides 8-10 times improvement over sequential. See "benches/flatmap.rs" or other variants.

Applied optimization is through using a binary heap to sort a sorted list of vectors. Further, we take advantage of composed push-and-pop operation to further improve the speed.

Even in the large-output benchmarks, where writing the output overweighs the computation, orx-parallel still outperforms.

In the special case where the work to be done as well as the size of the output are both insignificant, sequential is 70 times faster than rayon. While orx-parallel is only 30% slower when there are too few elements and 3 times faster when the input is longer. All together, new implementation seems to be very robust in different use cases.